requirements.txt
streamlit==1.29.0
pytz~=2023.3.post1
python-dotenv~=1.0.0
